change unit test file and class name
